Added DB interaction messages

+++ journal/src/modules/daybook/views/EntryView.vue
@@ -46,7 +46,7 @@
 import {defineAsyncComponent} from 'vue';
 import {mapGetters, mapActions} from 'vuex';
 import getDayMonthYear from '../helpers/getDayMonthYear.js';
-
+import Swal from 'sweetalert2';
 
 export default {
   name: "EntryView.vue",
@@ -101,16 +101,36 @@ export default {
       this.entry = entry;
     },
     async saveEntry() {
+      new Swal({
+        title: 'Espere por favor',
+        allowOutsideClick: false,
+      });
+      Swal.showLoading();
       if(this.entry.id) {
         await this.updateEntry(this.entry);
       } else {
         const id = await this.createEntry(this.entry);
         this.$router.push({name: 'entry', params: {id}});
       }
+      Swal.fire('Guardado', 'Entrada registrada con éxito', 'success');
     },
     async onDeleteEntry() {
-      await this.deleteEntry(this.entry.id);
-      this.$router.push({name: 'no-entry'});
+      const {isConfirmed} = await Swal.fire({
+        title: '¿Estás seguro?',
+        text: 'Una vez borrado, no se puede recuperar',
+        showDenyButton: true,
+        confirmButtonText: 'Si, estoy seguro'
+      });
+      if(isConfirmed) {
+        new Swal({
+          title: 'Espere por favor',
+          allowOutsideClick: false,
+        });
+        Swal.showLoading();
+        await this.deleteEntry(this.entry.id);
+        this.$router.push({name: 'no-entry'});
+        Swal.fire('Eliminado', '', 'Success');
+      }
     }
   },
   created() {
